Final.
Mudança de tipo para tamanho em sobremesa.
-Função excluir
-Algumas correções nos prints
:)

+++ AtividadePratica2/src/Alimento.java
@@ -19,7 +19,7 @@ public class Alimento {
         return preco;
     }
 
-    public void setPreco(double preco) {
+    public void setPreco(float preco) {
         this.preco = preco;
     }
